Added States and refactored code

+++ src/main/java/edu/neu/csye7374/strategypattern/CashPaymentStrategy.java
@@ -2,16 +2,16 @@ package edu.neu.csye7374.strategypattern;
 
 public class CashPaymentStrategy implements PaymentStrategy{
 
-    int paymentAmount;
-    int tipAmount;
+    double paymentAmount;
+    double tipAmount;
 
-    public CashPaymentStrategy(int paymentAmount, int tipAmount) {
+    public CashPaymentStrategy(double paymentAmount, double tipAmount) {
         this.paymentAmount = paymentAmount;
         this.tipAmount = tipAmount;
     }
 
     @Override
-    public void processPayment(int totalBillAmount) {
+    public void processPayment(double totalBillAmount) {
         System.out.println("CASH PAYMENT DETAILS");
         System.out.println("Total Bill Amount " + totalBillAmount);
         System.out.println("Tip Amount " + this.tipAmount);
